+++ Plugin/Thunderstore/CHANGELOG.md
@@ -1,3 +1,8 @@
+## v 2.1.3
+- Fixed bug with driftwood throwing people off the map.
+- Fixed some animation desyncs/bugs with driftwood and redwood.
+- Added Stellar Sovereign (doesn't currently do anything).
+
 ## v 2.1.2 *bug fixes*
 - Fixed config option with whistle price not working.
 - Downgraded mod back to LL for important bug fixes, please let me know if anything doesn't work.